read extra certs and use cert cache
